+++ dgraph/cmd/root_ee.go
@@ -14,13 +14,13 @@ package cmd
 
 import (
 	acl "github.com/dgraph-io/dgraph/ee/acl/cmd"
-	"github.com/dgraph-io/dgraph/ee/cmd/restore"
+	"github.com/dgraph-io/dgraph/ee/cmd/backup"
 )
 
 func init() {
 	// subcommands already has the default subcommands, we append to EE ones to that.
 	subcommands = append(subcommands,
-		&restore.Restore,
+		&backup.Restore,
 		&acl.CmdAcl,
 	)
 }
